refactor(categories): move category utility functions to a dedicated utils file
