+++ src/app/dashboard/accounting-book/Records.tsx
@@ -9,8 +9,11 @@ import { useAccountMap } from '@/hooks/useAccountMap';
 import { useCategoryMap } from '@/hooks/useCategoryMap';
 import { useConfirm } from '@/hooks/useConfirmModal';
 import { deleteAccountingRecord } from '@/lib/api/accounting';
-import { getCategoryIconById, getCategoryLabelById } from '@/lib/categories';
 import { getAccountLabelById } from '@/lib/utils/account';
+import {
+  getCategoryIconById,
+  getCategoryLabelById,
+} from '@/lib/utils/categories';
 import { AccountingRecord } from '@/types/accounting';
 
 interface RecordsProps {
