+++ src/app/dashboard/search/GroupCardSection.tsx
@@ -3,9 +3,12 @@ import { PhotoProvider, PhotoView } from 'react-photo-view';
 
 import { Button } from '@/components/ui/button';
 import { highlightText } from '@/components/ui/highlight-text';
-import { getCategoryIconById, getCategoryLabelById } from '@/lib/categories';
 import { formatToShortDay } from '@/lib/format';
 import { getAccountLabelById } from '@/lib/utils/account';
+import {
+  getCategoryIconById,
+  getCategoryLabelById,
+} from '@/lib/utils/categories';
 import { Account } from '@/types/account';
 import { AccountingRecord } from '@/types/accounting';
 import { Category } from '@/types/category';
